made corrections

+++ src/admin/controllers/products.js
@@ -1,15 +1,21 @@
 /* eslint-disable linebreak-style */
 /* eslint-disable max-len */
 const adminModel = require('../models/products');
+const adminMiddleware = require('../middleware/checkPermission');
 
 async function viewRoles(req, res) {
     try {
-        const { id } = req.params;
-        const result = await adminModel.viewRoles(id);
-        if (result) {
-            res.status(200).json({ roles: result });
+        const roles = await adminMiddleware.fetchRoles(req, res);
+        if (roles.includes(1)) {
+            const { id } = req.params;
+            const result = await adminModel.viewRoles(id);
+            if (result) {
+                res.status(200).json({ roles: result });
+            } else {
+                res.status(404).json({ message: 'Something went wrong' });
+            }
         } else {
-            res.status(404).json({ message: 'Something went wrong' });
+            res.status(401).json({ error: 'You are not a admin' });
         }
     } catch (err) {
         res.status(500).json({ error: 'Internal server error' });
@@ -18,17 +24,22 @@ async function viewRoles(req, res) {
 
 async function addRoles(req, res) {
     try {
-        const { id } = req.params;
-        const { roleId } = req.body;
-        if (roleId > 0 && roleId < 5) {
-            const result = await adminModel.addRoles(id, roleId);
-            if (result) {
-                res.status(200).json({ message: 'Role added' });
+        const roles = await adminMiddleware.fetchRoles(req, res);
+        if (roles.includes(1)) {
+            const { id } = req.params;
+            const { roleId } = req.body;
+            if (roleId > 0 && roleId < 5) {
+                const result = await adminModel.addRoles(id, roleId);
+                if (result) {
+                    res.status(200).json({ message: 'Role added' });
+                } else {
+                    res.status(404).json({ message: 'Something went wrong' });
+                }
             } else {
-                res.status(404).json({ message: 'Something went wrong' });
+                res.status(404).json({ message: 'RoleId should be between 1 - 4' });
             }
         } else {
-            res.status(404).json({ message: 'RoleId should be between 1 - 4' });
+            res.status(401).json({ error: 'You are not a admin' });
         }
     } catch (err) {
         res.status(500).json({ error: 'Internal server error' });
@@ -37,17 +48,22 @@ async function addRoles(req, res) {
 
 async function removeRoles(req, res) {
     try {
-        const { id } = req.params;
-        const { roleId } = req.body;
-        if (roleId > 0 && roleId < 5) {
-            const result = await adminModel.removeRoles(id, roleId);
-            if (result) {
-                res.status(200).json({ message: 'Role removed' });
+        const roles = await adminMiddleware.fetchRoles(req, res);
+        if (roles.includes(1)) {
+            const { id } = req.params;
+            const { roleId } = req.body;
+            if (roleId > 0 && roleId < 5) {
+                const result = await adminModel.removeRoles(id, roleId);
+                if (result) {
+                    res.status(200).json({ message: 'Role removed' });
+                } else {
+                    res.status(404).json({ message: 'Something went wrong' });
+                }
             } else {
-                res.status(404).json({ message: 'Something went wrong' });
+                res.status(404).json({ message: 'RoleId should be between 1 - 4' });
             }
         } else {
-            res.status(404).json({ message: 'RoleId should be between 1 - 4' });
+            res.status(401).json({ error: 'You are not a admin' });
         }
     } catch (err) {
         res.status(500).json({ error: 'Internal server error' });
@@ -56,16 +72,21 @@ async function removeRoles(req, res) {
 
 async function addProduct(req, res) {
     try {
-        const {
-            name, sku, description, price, stock, maxLimitPerOrder, categoryId, discount, sellerId,
-        } = req.body;
-        const imageName = req.file.filename;
-        const imageLink = `/img/${imageName}`;
-        const result = await adminModel.addProduct(name, sku, description, price, stock, maxLimitPerOrder, categoryId, discount, sellerId, imageLink);
-        if (result) {
-            res.status(200).json({ message: 'Item added successfully' });
+        const roles = await adminMiddleware.fetchRoles(req, res);
+        if (roles.includes(1) || roles.includes(2)) {
+            const {
+                name, sku, description, price, stock, maxLimitPerOrder, categoryId, discount, sellerId,
+            } = req.body;
+            const imageName = req.file.filename;
+            const imageLink = `/img/${imageName}`;
+            const result = await adminModel.addProduct(name, sku, description, price, stock, maxLimitPerOrder, categoryId, discount, sellerId, imageLink);
+            if (result) {
+                res.status(200).json({ message: 'Item added successfully' });
+            } else {
+                res.status(404).json({ message: 'Something went wrong' });
+            }
         } else {
-            res.status(404).json({ message: 'Something went wrong' });
+            res.status(401).json({ error: 'You dont have access to Add Items' });
         }
     } catch (err) {
         res.status(500).json({ error: 'Internal server error' });
@@ -74,14 +95,19 @@ async function addProduct(req, res) {
 
 async function addBannerImage(req, res) {
     try {
-        const imageName = req.file.filename;
-        const imageTitle = req.body.title;
-        const imageLink = `/img/banner/${imageName}`;
-        const result = await adminModel.addBannerImage(imageTitle, imageLink);
-        if (result) {
-            res.status(200).json({ message: 'Image added successfully' });
+        const roles = await adminMiddleware.fetchRoles(req, res);
+        if (roles.includes(1)) {
+            const imageName = req.file.filename;
+            const imageTitle = req.body.title;
+            const imageLink = `/img/banner/${imageName}`;
+            const result = await adminModel.addBannerImage(imageTitle, imageLink);
+            if (result) {
+                res.status(200).json({ message: 'Image added successfully' });
+            } else {
+                res.status(404).json({ message: 'Something went wrong' });
+            }
         } else {
-            res.status(404).json({ message: 'Something went wrong' });
+            res.status(401).json({ error: 'You are not a admin' });
         }
     } catch (err) {
         res.status(500).json({ error: 'Internal server error' });
@@ -90,12 +116,17 @@ async function addBannerImage(req, res) {
 
 async function deleteBannerImage(req, res) {
     try {
-        const imageId = req.params.id;
-        const result = await adminModel.deleteBannerImage(imageId);
-        if (result) {
-            res.status(200).json({ message: 'Image Deleted successfully' });
+        const roles = await adminMiddleware.fetchRoles(req, res);
+        if (roles.includes(1)) {
+            const imageId = req.params.id;
+            const result = await adminModel.deleteBannerImage(imageId);
+            if (result) {
+                res.status(200).json({ message: 'Image Deleted successfully' });
+            } else {
+                res.status(404).json({ message: 'Something went wrong' });
+            }
         } else {
-            res.status(404).json({ message: 'Something went wrong' });
+            res.status(401).json({ error: 'You are not a admin' });
         }
     } catch (err) {
         res.status(500).json({ error: 'Internal server error' });
@@ -117,13 +148,18 @@ async function fetchBannerImage(req, res) {
 
 async function updateProduct(req, res) {
     try {
-        const productId = req.params.id;
-        const updatedDetails = req.body;
-        const result = await adminModel.updateProduct(updatedDetails, productId);
-        if (result) {
-            res.status(200).json({ message: 'Item updated successfully' });
+        const roles = await adminMiddleware.fetchRoles(req, res);
+        if (roles.includes(1) || roles.includes(3)) {
+            const productId = req.params.id;
+            const updatedDetails = req.body;
+            const result = await adminModel.updateProduct(updatedDetails, productId);
+            if (result) {
+                res.status(200).json({ message: 'Item updated successfully' });
+            } else {
+                res.status(404).json({ message: 'Something went wrong' });
+            }
         } else {
-            res.status(404).json({ message: 'Something went wrong' });
+            res.status(401).json({ error: 'You dont have access to update' });
         }
     } catch (err) {
         res.status(500).json({ error: 'Internal server error' });
@@ -132,12 +168,17 @@ async function updateProduct(req, res) {
 
 async function deleteProduct(req, res) {
     try {
-        const productId = req.params.id;
-        const result = await adminModel.deleteProduct(productId);
-        if (result) {
-            res.status(200).json({ message: 'Item deleted successfully' });
+        const roles = await adminMiddleware.fetchRoles(req, res);
+        if (roles.includes(1) || roles.includes(4)) {
+            const productId = req.params.id;
+            const result = await adminModel.deleteProduct(productId);
+            if (result) {
+                res.status(200).json({ message: 'Item deleted successfully' });
+            } else {
+                res.status(404).json({ message: 'Something went wrong' });
+            }
         } else {
-            res.status(404).json({ message: 'Something went wrong' });
+            res.status(401).json({ error: 'You dont have access to delete' });
         }
     } catch (err) {
         res.status(500).json({ error: 'Internal server error' });
@@ -146,12 +187,16 @@ async function deleteProduct(req, res) {
 
 async function fetchProducts(req, res) {
     try {
-        const result = await adminModel.fetchProducts();
-
-        if (result.length > 0) {
-            res.status(200).json({ products: result });
+        const roles = await adminMiddleware.fetchRoles(req, res);
+        if (roles.includes(1) || roles.includes(2) || roles.includes(3) || roles.includes(4)) {
+            const result = await adminModel.fetchProducts();
+            if (result.length > 0) {
+                res.status(200).json({ products: result });
+            } else {
+                res.status(404).json({ message: 'No products to display' });
+            }
         } else {
-            res.status(404).json({ message: 'No products to display' });
+            res.status(401).json({ error: 'You dont have access' });
         }
     } catch (err) {
         res.status(500).json({ error: 'Internal server error' });
